add rest endpoints for customer crud

+++ src/main/java/example/guesthousecustomerservice/services/impl/CustomerServiceImpl.java
@@ -1,15 +1,15 @@
 package example.guesthousecustomerservice.services.impl;
 
 
-
-
-
 import example.guesthousecustomerservice.clients.BookingServiceClient;
 import example.guesthousecustomerservice.dtos.CustomerDTO;
 import example.guesthousecustomerservice.exceptions.CustomerHasActiveBookingsException;
 import example.guesthousecustomerservice.models.Customer;
 import example.guesthousecustomerservice.repositories.CustomerRepository;
 import example.guesthousecustomerservice.services.CustomerService;
+
+import java.util.NoSuchElementException;
+
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -44,19 +44,31 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public void save(CustomerDTO customerDTO) {
+    public CustomerDTO save(CustomerDTO customerDTO) {
         Customer customer = new Customer(customerDTO.getName());
-        customer.setId(customerDTO.getId());
-        customerRepository.save(customer);
-
+        Customer saved = customerRepository.save(customer);
+        return new CustomerDTO(saved.getName(), saved.getId());
     }
+
+    @Override
+    public CustomerDTO update(Long id, CustomerDTO customerDTO) {
+        Customer customer = customerRepository.findById(id).orElseThrow();
+        customer.setName(customerDTO.getName());
+        Customer saved = customerRepository.save(customer);
+        return new CustomerDTO(saved.getName(), saved.getId());
+    }
+
     @Override
     public void delete(Long id) {
+        if (!customerRepository.existsById(id)) {
+            throw new NoSuchElementException();
+        }
         if (bookingServiceClient.hasActiveBookings(id)) {
             throw new CustomerHasActiveBookingsException("Can't remove customer with active booking!");
         }
         customerRepository.deleteById(id);
     }
-
 }
+
+
